Date & Dish carousel with draft visual
