+++ client/src/index.css
@@ -12,7 +12,7 @@
 }
 
 .custom-bg {
-  @apply h-screen bg-cover bg-center relative;
+  @apply bg-cover bg-center relative;
   background-image: url('https://res.cloudinary.com/dspxn4ees/image/upload/v1738147201/calendar-background-image.jpg');
   padding-top: 25px;
 }
